# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Koyomi Gal 后端 API",
        "title": "Koyomi Gal API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/admin/galgames": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查询全部状态（pending/published/rejected/hidden）的 Galgame；需要 galgame:review 权限",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "管理端查询 Galgame",
                "operationId": "listAdminGalgames",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "状态过滤：0 待审，1 已发布，2 已拒绝，3 已隐藏",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标题或别名关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "latest",
                        "description": "排序：latest、oldest、rating、favorite、popular",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量，最大 100",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Galgame 列表",
                        "schema": {
                            "$ref": "#/definitions/dto.GalgameListResponse"
                        }
                    },
                    "400": {
                        "description": "查询参数格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "没有执行该操作的权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "查询 Galgame 失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/galgames/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "按 ID 返回任意状态 Galgame 详情；需要 galgame:review 权限",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "管理端查看 Galgame 详情",
                "operationId": "getAdminGalgame",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Galgame ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Galgame 详情",
                        "schema": {
                            "$ref": "#/definitions/dto.GalgameDataResponse"
                        }
                    },
                    "400": {
                        "description": "Galgame ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "没有执行该操作的权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Galgame 不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "查询 Galgame 失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/developers": {
            "get": {
                "description": "返回全部开发商",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "developers"
                ],
                "summary": "查看开发商列表",
                "operationId": "listDevelopers",
                "responses": {
                    "200": {
                        "description": "开发商列表",
                        "schema": {
                            "$ref": "#/definitions/dto.DeveloperListResponse"
                        }
                    },
                    "500": {
                        "description": "查询开发商失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建 Galgame 开发商",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "developers"
                ],
                "summary": "创建开发商",
                "operationId": "createDeveloper",
                "parameters": [
                    {
                        "description": "创建开发商请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateDeveloperRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/dto.DeveloperDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "没有执行该操作的权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "slug 已存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "创建开发商失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/tags": {
            "get": {
                "description": "返回全部 Galgame Tag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "查看 Tag 列表",
                "operationId": "listTags",
                "responses": {
                    "200": {
                        "description": "Tag 列表",
                        "schema": {
                            "$ref": "#/definitions/dto.TagListResponse"
                        }
                    },
                    "500": {
                        "description": "查询 Tag 失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建 Galgame Tag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "创建 Tag",
                "operationId": "createTag",
                "parameters": [
                    {
                        "description": "创建 Tag 请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/dto.TagDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "没有执行该操作的权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "名称或 slug 已存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "创建 Tag 失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateDeveloperRequest": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "游戏开发商"
                },
                "logo_url": {
                    "type": "string",
                    "example": "https://example.com/logo.png"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "YUZUSOFT"
                },
                "original_name": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "ゆずソフト"
                },
                "slug": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "yuzusoft"
                },
                "website": {
                    "type": "string",
                    "example": "https://www.yuzu-soft.com"
                }
            }
        },
...
        "dto.DeveloperListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DeveloperResponse"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "dto.GalgameDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/dto.GalgameResponse"
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "dto.GalgameListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/dto.GalgameListData"
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 233
                },
                "msg": {
                    "type": "string",
                    "example": "业务错误"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "传入 \"Bearer {access_token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}